[BUG-124] prevent leaking of MySQL schemas
MySQL does not use the Catalog abstraction, but each MySQL database is a
JDBC schema. When creating the sql meta data the former SchemaProvider did
not check which one is the currently active schema. This is fixed by querying
the current active schema, and filtering meta data accordingly. MySQL specific
internal schemas are forbidden by default. There are now two distinct
SchemaProvider, one for MySQL and one for H2. Conversion from the JOOQ
model to the asio model is extracted to the new SqlSchemaBuilder class.
The providers are now located in the lair module.
For testing purposes an explorative test, which shows the proprietary
MySQL behavior and a d2r mapping for a local MySQL instance on JOSIE are
added.