File/Path test tools
Move TransientFile to test module. References to keystore removed from class.

Add JUnit rule to create temporary folders. Mimics functionality of built-in
TemporaryFolder, but with a simplified API and uses JDK7 Path/Files internally.

Add FindFiles visitor that uses a PathMatcher to decide if a visited file should be
included in the collection of found files.

Add hamcrest matchers for JDK7 Paths