Support for configuring schema and identifier
ContainerSettings.Datasource now allows to set the name of the schema in
the backing database independently from the container name. In d2rq
mappings, this can be set by adding a jdbc:schema proeprty to the database
configuration.

D2rqSpec uses d2r:Server resource uri as container identifier. For vph
deployments, where that resource is named '' this means, that the base
resource uri is used as identifier.