Helper for spring classpath scanning
Add SpringAutoFactory marker interface.
Needed, as google-auto-factory is not able to annotate generated types.
Therefore @Component cannot be applied to generated factories and they are
not eligible for classpath scanning. If annotated with this marker
interface, they can be included in classpath scanning by adding an
inclusion filter targeting the marker interface to @ComponentScan.

Add ExplicitWiring annotation, that may be added to beans, that have to be
annotated as @Component, but must not be included in classpath scanning.