Simplify spring security components
Simplify implementations and improve testability of authentication
implementation.

The HTTP method restriction is now applied in a separate filter, after
authentication is complete.

The delegated credentials header name is now conigurable and credentials
are now stored in the Authentication#getDetails().

Implement uri auth as variant of the PreAuthenticationFilter. The
principal name is extracted from the request uri. No credentials are
provided.

Add RoleUserService that holds a fixed mapping of Role names to users, where
the user name is equal to the role name. Each user has the authorities
granted to its corresponding role. Matching of principal/username is case
insensitive.

Add simple redirect filter for explorer page requests.