Global datasource feature
Allow overriding the jdbc connection information with a global setting. To
enable multi-tenancy, the hikari pool and d2r SQLConnection are configured
to switch to a container specific catalog in initialization. SQLConnection
now also includes jdbc connection properties.

The integration test configurations are now merged into a single
application.yml file.
Add an integration mapping suitable for h2 and mysql and use a factory class
for integration databases, which is able to detect if a mysql database is
available. VphIntegrationSuite uses a mysql database if possible, as does
the Runner.
Cut long identifiers in gui test schema as MySql limits identifier names
to a maximum length of 64 characters. Test schema for gui exceeded this value.
Shortened to be exacly 64 characters long.

Context now reports profiles and settings on boot. D2rqNestAssembler no
longer logs in ctor.