Emit events on sql executions
Add a mysql StatementInterceptor, which emites SqlEvents describing the
executed sql commands. Add a 'mysql' profile to the asio default config
with default settings for statement caching and the interceptor. The MySql
connector is now a compile time dependency.

Move H2SchemaServiceTest and MysqlSchemaExploration to correct package.

Filter events in integration tests while receiving, to improve resilience
of test implementations.