Fix lost mysql-user on redeployment
If a container is redeployed, while multi-tenancy is active the assembly
of the new container instance is interleaved with the destruction of the
former instance. That includes the mysql user management and leads to the
user being created first and then being dropped immediately. Startup of
the redeployed container then fails.

Fixed by adding a @PostConstruct hook in NestBluePrint, which ensures the
mysql user is present (recreating it if necessary). Add a test case to
detect regressions.