Validate and refresh pooled d2rq models
D2rqModelAllocator now ensures a pooled model is live during reallocation.
A model is live if it is (a) not closed and (b) the connection to the
backing rdbms is valid. If a model is not live, it is disposed and
replaced by a fresh one.
The expiration policy of the PooledD2rqFactory has been changed to a short
expiration period of less than a minute, to ensure that models are
frequently validated and replaced fast if they are failing.

The d2rq submodule now references a commit in origin with changes to
GraphD2RQ required for the model validation.

This fixes https://github.com/pyranja/asio/issues/1 .