Minimize explicit values in server.yml
server.yml is exposed to the user as runtime configuration file. Instead
of explicitly overriding many values here, comment all properties that are
equal to the internal defaults (from application.yml), to avoid breaking
default installations on updates and at the same time show options to the
deployer.

Default password changed to be randomly set (this breaks out of the box
administration) to prevent publicly access to out of the box
installations.