Fail deployment if d2rq configuration is corrupt
PooledD2rqFactory now tries to create a single D2RQ model on instantiation
to detect invalid d2rq configuration immediately. The thrown exception
will cancel the deployment of the dataset.

This fixes #4, where the stormpot pool tried to repeatedly allocate new
d2rq models. If the configuration is corrupt, model allocation is bound to
fail, but as the pool is filled in a dedicated thread exceptions are
swallowed and it simply continues to try and fail forever.